More Default Commands, Shooter/Chassis Commands
This commit adds default commands to the Intake, Climber, and Shooter
subsystems; and it adds the necessary #includes to the relevant
subsystem file.
In addition, the ChassisTankDrive is updated to hopefully maintain left
and right continuity when driving is reversed.
A command to feed the shooters was added, and some commands have been
assigned buttons on the coDriver controller.