Auto choosing revamp; Removal of auto commands from dashboard
The auto SendableChooser has been changed so it is no longer a pointer
and it now accepts a string rather than an int as a result of a weird
error regarding the SendableChooser.h file. AutoCommand() was revamped
to reflect this change, now accepting inputs such as "auto0" "auto1"
etc.
Autonomous command buttons were removed from the SmartDashboard.